# backend/openapi.yaml
openapi: 3.0.0
info:
  title: CliffBnb Public User API
  description: >
    User-facing (guest, host, admin) API contract for CliffBnb.  

    All entity types, fields, and parameter naming strictly match Zod schemas
    and UX datamap specs.  

    All IDs are strings. All field names use snake_case. All endpoints reference
    exact Zod schema fields.
  version: 1.0.0
servers:
  - url: http://localhost:3000
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    LocationParam:
      name: location
      in: query
      description: Filter by location (city, region, country string).
      schema:
        type: string
    AmenitiesParam:
      name: amenities
      in: query
      description: Comma-separated list of amenity slugs.
      schema:
        type: string
    StartDateParam:
      name: start_date
      in: query
      description: Start of booking date range (YYYY-MM-DD).
      schema:
        type: string
        format: date
    EndDateParam:
      name: end_date
      in: query
      description: End of booking date range (YYYY-MM-DD).
      schema:
        type: string
        format: date
    AdultsParam:
      name: adults
      in: query
      description: Number of adults.
      schema:
        type: integer
    ChildrenParam:
      name: children
      in: query
      description: Number of children.
      schema:
        type: integer
    InfantsParam:
      name: infants
      in: query
      description: Number of infants.
      schema:
        type: integer
    PriceMinParam:
      name: price_min
      in: query
      description: Minimum price per night.
      schema:
        type: integer
    PriceMaxParam:
      name: price_max
      in: query
      description: Maximum price per night.
      schema:
        type: integer
    SortByParam:
      name: sort_by
      in: query
      description: Sort field (created_at, price_per_night, average_rating, review_count).
      schema:
        type: string
        enum:
          - created_at
          - price_per_night
          - average_rating
          - review_count
    SortOrderParam:
      name: sort_order
      in: query
      description: Sort order asc/desc.
      schema:
        type: string
        enum:
          - asc
          - desc
    PageParam:
      name: page
      in: query
      description: Page number (1-based).
      schema:
        type: integer
        minimum: 1
        default: 1
...
    ViewModeParam:
      name: view_mode
      in: query
      description: Search view type (list or map).
      schema:
        type: string
        enum:
          - list
          - map
    NeLatParam:
      name: ne_lat
      in: query
      description: Map northeast latitude bound.
      schema:
        type: number
    NeLngParam:
      name: ne_lng
      in: query
      description: Map northeast longitude bound.
      schema:
        type: number
    SwLatParam:
      name: sw_lat
      in: query
      description: Map southwest latitude bound.
      schema:
        type: number
    SwLngParam:
      name: sw_lng
      in: query
      description: Map southwest longitude bound.
      schema:
        type: number
  schemas:
    User:
      type: object
      required:
        - user_id
        - email
        - name
        - password_hash
        - role
        - is_email_confirmed
        - has_unread_messages
        - has_unread_notifications
        - created_at
        - updated_at
      properties:
        user_id:
          type: string
        email:
          type: string
          format: email
        name:
          type: string
        password_hash:
          type: string
        role:
          type: string
        profile_photo_url:
          type: string
          nullable: true
        contact_info:
          type: string
          nullable: true
        host_bio:
          type: string
          nullable: true
        is_email_confirmed:
          type: boolean
        email_confirmation_token:
          type: string
          nullable: true
        password_reset_token:
          type: string
          nullable: true
        has_unread_messages:
          type: boolean
        has_unread_notifications:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreateUserInput:
      type: object
      required:
        - email
        - name
        - password_hash
        - role
      properties:
        email:
          type: string
          format: email
        name:
          type: string
        password_hash:
          type: string
        role:
          type: string
        profile_photo_url:
          type: string
          nullable: true
        contact_info:
          type: string
          nullable: true
        host_bio:
          type: string
          nullable: true
    UpdateUserInput:
      type: object
      required:
        - user_id
      properties:
        user_id:
          type: string
        email:
          type: string
          format: email
        name:
          type: string
        password_hash:
          type: string
        role:
          type: string
        profile_photo_url:
          type: string
          nullable: true
        contact_info:
          type: string
          nullable: true
        host_bio:
          type: string
          nullable: true
        is_email_confirmed:
          type: boolean
        email_confirmation_token:
          type: string
          nullable: true
        password_reset_token:
          type: string
          nullable: true
        has_unread_messages:
          type: boolean
        has_unread_notifications:
          type: boolean
    UserList:
      type: array
      items:
        $ref: "#/components/schemas/User"
    AuthToken:
      type: object
      required:
        - token
        - user
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
    AuthRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
    EmailRequest:
      type: object
      required:
        - email
      properties:
        email:
          type: string
          format: email
    ResetPasswordInput:
      type: object
      required:
        - token
        - password
      properties:
        token:
          type: string
        password:
          type: string
          minLength: 8
...
    CreateVillaInput:
      type: object
      required:
        - host_user_id
        - name
        - location
        - lat
        - lng
        - description
        - amenities
        - price_per_night
        - cancellation_policy
      properties:
        host_user_id:
          type: string
        name:
          type: string
        subtitle:
          type: string
          nullable: true
        location:
          type: string
        lat:
          type: number
        lng:
          type: number
        address:
          type: string
          nullable: true
        description:
          type: string
        house_rules:
          type: string
          nullable: true
        special_notes:
          type: string
          nullable: true
        amenities:
          type: string
        price_per_night:
          type: number
        cleaning_fee:
          type: number
        service_fee:
          type: number
        minimum_stay_nights:
          type: integer
        cancellation_policy:
          type: string
        status:
          type: string
        occupancy:
          type: integer
...
    VillaList:
      type: array
      items:
        $ref: "#/components/schemas/Villa"
...
paths:
  /auth/signup:
    post:
      summary: Register a new user
      operationId: signupUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserInput"
      responses:
        "201":
          description: User created, confirmation link sent, returns JWT token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthToken"
  /auth/login:
    post:
      summary: Sign in with email and password
      operationId: loginUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthRequest"
      responses:
        "200":
          description: Returns JWT token and user details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthToken"
  /auth/confirm-email:
    post:
      summary: Confirm user email by token
      operationId: confirmEmail
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - token
              properties:
                token:
                  type: string
      responses:
        "200":
          description: Email confirmed successfully, returns JWT token & user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthToken"
  /auth/forgot-password:
    post:
      summary: Send password reset link to user email
      operationId: sendPasswordReset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmailRequest"
      responses:
        "200":
          description: Password reset link sent
  /auth/reset-password:
    post:
      summary: Reset user password with token from email
      operationId: resetPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetPasswordInput"
      responses:
        "200":
          description: Password changed, returns new JWT token & user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthToken"
...
  /users/{user_id}:
    get:
      summary: Get details for a public user profile
      operationId: getUser
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User object
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    put:
      summary: Update current user profile (full)
      operationId: updateUser
      security:
        - bearerAuth: []
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserInput"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    patch:
      summary: Patch current user fields
      operationId: patchUser
      security:
        - bearerAuth: []
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserInput"
      responses:
        "200":
          description: Patched user profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/me:
    get:
      summary: Get the authenticated user's profile & dashboard info
      operationId: getOwnUser
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user object
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users:
    get:
      summary: Search/list of users (admin, moderation)
      operationId: searchUsers
      security:
        - bearerAuth: []
      parameters:
        - name: query
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
        - name: offset
          in: query
          schema:
            type: integer
            minimum: 0
        - name: role
          in: query
          schema:
            type: string
        - name: is_email_confirmed
          in: query
          schema:
            type: boolean
        - name: sort_by
          in: query
          schema:
            type: string
            enum:
              - name
              - created_at
              - email
        - name: sort_order
          in: query
          schema:
            type: string
            enum:
              - asc
              - desc
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserList"
  /villas:
    get:
      summary: Search/cliff villa listings with filters (homepage, search, map)
      operationId: searchVillas
      parameters:
        - $ref: "#/components/parameters/LocationParam"
        - $ref: "#/components/parameters/AmenitiesParam"
        - $ref: "#/components/parameters/StartDateParam"
        - $ref: "#/components/parameters/EndDateParam"
        - $ref: "#/components/parameters/AdultsParam"
        - $ref: "#/components/parameters/ChildrenParam"
        - $ref: "#/components/parameters/InfantsParam"
        - $ref: "#/components/parameters/PriceMinParam"
        - $ref: "#/components/parameters/PriceMaxParam"
        - $ref: "#/components/parameters/SortByParam"
        - $ref: "#/components/parameters/SortOrderParam"
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/ViewModeParam"
        - $ref: "#/components/parameters/NeLatParam"
        - $ref: "#/components/parameters/NeLngParam"
        - $ref: "#/components/parameters/SwLatParam"
        - $ref: "#/components/parameters/SwLngParam"
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
        - name: status
          in: query
          description: Filter published/unpublished for admin/host
          schema:
            type: string
      responses:
        "200":
          description: Villa listing matching filters, pagination
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VillaList"
    post:
      summary: Host creates a new villa
      operationId: createVilla
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateVillaInput"
      responses:
        "201":
          description: Created villa
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Villa"
...